+++ text/python/text.py
@@ -156,7 +156,7 @@ class Text:
         if not self.okay:
             return
 
-        self.led_strip.set_frame_duration(1000.0 / config.TEXT_FRAME_RATE)
+        self.led_strip.set_frame_duration(int(1000.0 / config.TEXT_FRAME_RATE))
 
     def set_new_text(self, text):
         text = '   ' + text
